+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.7.7] - 2025-11-23
+
 ### Added
 
 - GTCEU Recipes now will show duration, level and every consumption